@@ -1,15 +1,11 @@
 import { Component, OnInit, TemplateRef, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import {
-    TableCellType,
-    TableChangeEvent,
-    TableColumn,
-    TableColumnFilter,
-    TableConfiguration, TableFilterType
-} from '../../../shared/hazlenut/core-table';
+import { Role } from '../../../shared/enums/role.enum';
+import { TableCellType, TableChangeEvent, TableColumn, TableColumnFilter, TableConfiguration, TableFilterType } from '../../../shared/hazlenut/core-table';
 import { BrowseResponse } from '../../../shared/hazlenut/hazelnut-common/models';
 import { BusinessArea } from '../../../shared/interfaces/bussiness-area.interface';
+import { AuthService } from '../../../shared/services/auth.service';
 import { BusinessAreaService } from '../../../shared/services/data/business-area.service';
 import { NotificationService } from '../../../shared/services/notification.service';
 import { SelectedAreaService } from '../../../shared/services/storage/selected-area.service';
@@ -22,22 +18,21 @@ import { SelectedAreaService } from '../../../shared/services/storage/selected-a
 
 /**
  * Business area list with title table and loader
- */
-export class BusinessAreaListComponent implements OnInit {
-    // if expanded content is not used, then we need to remove it
-    @ViewChild('expandedContent') public expandedContent: TemplateRef<any>;
-    @ViewChild('navigationToTasksColumn') public navigationToTasksColumn: TemplateRef<any>;
-    private isInitialized = false;
+ */ export class BusinessAreaListComponent implements OnInit {
+    // If expanded content is not used, then we need to remove it
+    @ViewChild('expandedContent', {static: true}) public expandedContent: TemplateRef<any>;
+    @ViewChild('navigationToTasksColumn', {static: true}) public navigationToTasksColumn: TemplateRef<any>;
     public loading = false;
     public config: TableConfiguration;
     public data: BrowseResponse<BusinessArea> = new BrowseResponse<BusinessArea>();
+    private isInitialized = false;
 
     public constructor(private readonly translateService: TranslateService,
                        private readonly router: Router,
                        private readonly businessAreaService: BusinessAreaService,
                        private readonly selectedAreaService: SelectedAreaService,
                        private readonly notificationService: NotificationService,
-    ) {
+                       private readonly authService: AuthService) {
     }
 
     /**
@@ -92,14 +87,27 @@ export class BusinessAreaListComponent implements OnInit {
      */
     public setTableData(tableChangeEvent?: TableChangeEvent): void {
         this.loading = true;
-        this.businessAreaService.browseBusinessAreas(tableChangeEvent).subscribe((data) => {
-            this.data = data;
-            this.loading = false;
-            this.isInitialized = true;
-        }, () => {
-            this.loading = false;
-            this.notificationService.openErrorNotification('error.api');
-        });
+        this.businessAreaService.browseBusinessAreas(tableChangeEvent)
+            .subscribe((data) => {
+                this.data = data;
+                this.loading = false;
+                this.isInitialized = true;
+            }, () => {
+                this.loading = false;
+                this.notificationService.openErrorNotification('error.api');
+            });
+    }
+
+    public allowTaskListButton(): boolean {
+        return this.hasRoleReadTask() || this.hasRoleReadTaskInAssignProject();
+    }
+
+    private hasRoleReadTask(): boolean {
+        return this.authService.hasRole(Role.RoleReadTask);
+    }
+
+    private hasRoleReadTaskInAssignProject(): boolean {
+        return this.authService.hasRole(Role.RoleReadTaskInAssignProject);
     }
 
 }
