@@ -1,8 +1,10 @@
 import { Component, OnInit, TemplateRef, ViewChild } from '@angular/core';
+import { Role } from '../../../shared/enums/role.enum';
 import { TableCellType, TableColumn, TableConfiguration } from '../../../shared/hazlenut/core-table';
 import { BrowseResponse } from '../../../shared/hazlenut/hazelnut-common/models';
 import { FileManager } from '../../../shared/hazlenut/hazelnut-common/utils/file-manager';
 import { Report } from '../../../shared/interfaces/report.interface';
+import { AuthService } from '../../../shared/services/auth.service';
 import { ReportService } from '../../../shared/services/data/report.service';
 import { NotificationService } from '../../../shared/services/notification.service';
 import { ProjectEventService } from '../../../shared/services/storage/project-event.service';
@@ -14,18 +16,17 @@ import { GetFileNameFromContentDisposition } from '../../../shared/utils/headers
     styleUrls: ['./report-list.component.scss']
 })
 export class ReportListComponent implements OnInit {
-    @ViewChild('descriptionColumn') public descriptionColumn: TemplateRef<any>;
-    @ViewChild('actionColumn') public actionColumn: TemplateRef<any>;
+    @ViewChild('descriptionColumn', {static: true}) public descriptionColumn: TemplateRef<any>;
+    @ViewChild('actionColumn', {static: true}) public actionColumn: TemplateRef<any>;
 
     public loading = false;
     public config: TableConfiguration;
     public data = new BrowseResponse<Report>();
 
-    public constructor(
-        private readonly notificationService: NotificationService,
-        private readonly reportService: ReportService,
-        private readonly projectEventService: ProjectEventService,
-    ) {
+    public constructor(private readonly notificationService: NotificationService,
+                       private readonly reportService: ReportService,
+                       private readonly projectEventService: ProjectEventService,
+                       private readonly authService: AuthService) {
     }
 
     /**
@@ -58,37 +59,63 @@ export class ReportListComponent implements OnInit {
     }
 
     /**
-     * Set report table data from API
-     */
-    private setTableData() {
-        this.loading = true;
-        this.reportService.getAllReports().subscribe((data) => {
-            this.data.content = data;
-            this.data.totalElements = data.length;
-            this.loading = false;
-        }, () => {
-            this.loading = false;
-            this.notificationService.openErrorNotification('error.api');
-        });
-    }
-
-    /**
      * Export excel report from API
      * @param reportId
      */
     public export(reportId: number) {
         this.loading = true;
-        this.reportService.exportReport(this.projectEventService.instant.id, reportId).subscribe((response) => {
-            new FileManager().saveFile(
-                GetFileNameFromContentDisposition(response.headers.get('Content-Disposition')),
-                response.body,
-                'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
-            );
-            this.loading = false;
-        }, () => {
-            this.notificationService.openErrorNotification('error.api');
-            this.loading = false;
-        });
+        this.reportService.exportReport(this.projectEventService.instant.id, reportId)
+            .subscribe((response) => {
+                new FileManager().saveFile(
+                    GetFileNameFromContentDisposition(response.headers.get('Content-Disposition')),
+                    response.body,
+                    'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+                );
+                this.loading = false;
+            }, () => {
+                this.notificationService.openErrorNotification('error.api');
+                this.loading = false;
+            });
+    }
+
+    public allowButtonRedFlagList(): boolean {
+        return this.hasRoleExportReportRedFlagList() || this.hasRoleExportReportRedFlagListInAssignProject();
+    }
+
+    public allowButtonToDoList(): boolean {
+        return this.hasRoleExportReportToDoList() || this.hasRoleExportReportToDoListInAssignProject();
+    }
+
+    private hasRoleExportReportToDoList(): boolean {
+        return this.authService.hasRole(Role.RoleExportReportToDoList);
+    }
+
+    private hasRoleExportReportToDoListInAssignProject(): boolean {
+        return this.authService.hasRole(Role.RoleExportReportToDoListInAssignProject);
+    }
+
+    private hasRoleExportReportRedFlagList(): boolean {
+        return this.authService.hasRole(Role.RoleExportReportRedFlagList);
+    }
+
+    private hasRoleExportReportRedFlagListInAssignProject(): boolean {
+        return this.authService.hasRole(Role.RoleExportReportRedFlagListInAssignProject);
+    }
+
+    /**
+     * Set report table data from API
+     */
+    private setTableData() {
+        this.loading = true;
+        this.reportService.getAllReports()
+            .subscribe((data) => {
+                this.data.content = data;
+                this.data.totalElements = data.length;
+                this.loading = false;
+            }, () => {
+                this.loading = false;
+                this.notificationService.openErrorNotification('error.api');
+            });
     }
 
 }
