@@ -2,8 +2,11 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { tap } from 'rxjs/internal/operators/tap';
+import { Role } from '../../../shared/enums/role.enum';
 import { Regex } from '../../../shared/hazlenut/hazelnut-common/regex/regex';
 import { TaskComment, TaskCommentResponse } from '../../../shared/interfaces/task-comment.interface';
+import { AuthService } from '../../../shared/services/auth.service';
+import { ImagesService } from '../../../shared/services/data/images.service';
 import { TaskService } from '../../../shared/services/data/task.service';
 import { NotificationService } from '../../../shared/services/notification.service';
 import { ProjectEventService } from '../../../shared/services/storage/project-event.service';
@@ -16,23 +19,26 @@ import { TaskFormComponent } from '../task-form/task-form.component';
     styleUrls: ['./task-edit.component.scss']
 })
 export class TaskEditComponent implements OnInit {
-    @ViewChild(TaskFormComponent) public taskForm: TaskFormComponent;
+    @ViewChild(TaskFormComponent, {static: true}) public taskForm: TaskFormComponent;
     public formData = null;
     public addCommentForm: FormGroup;
     public comments: TaskCommentResponse[] = [];
     public loading = false;
     public notOnlyWhiteCharactersPattern = Regex.notOnlyWhiteCharactersPattern;
+    public attachmentFormat = '';
+    public attachmentFileName = '';
+    public attachmentPathName = '';
     private taskId: number;
 
-    public constructor(
-        private readonly router: Router,
-        private readonly taskCommentService: TaskCommentService,
-        private readonly notificationService: NotificationService,
-        private readonly activatedRoute: ActivatedRoute,
-        private readonly taskService: TaskService,
-        private readonly formBuilder: FormBuilder,
-        public readonly projectEventService: ProjectEventService,
-    ) {
+    public constructor(private readonly router: Router,
+                       private readonly taskCommentService: TaskCommentService,
+                       private readonly notificationService: NotificationService,
+                       private readonly activatedRoute: ActivatedRoute,
+                       private readonly taskService: TaskService,
+                       private readonly formBuilder: FormBuilder,
+                       private readonly imagesService: ImagesService,
+                       public readonly projectEventService: ProjectEventService,
+                       private readonly authService: AuthService) {
     }
 
     public ngOnInit(): void {
@@ -41,7 +47,11 @@ export class TaskEditComponent implements OnInit {
             this.getAllComments();
         });
         this.addCommentForm = this.formBuilder.group({
-            newComment: ['', Validators.required]
+            newComment: [
+                '',
+                Validators.required
+            ],
+            attachment: ['']
         });
     }
 
@@ -51,8 +61,8 @@ export class TaskEditComponent implements OnInit {
 
     public onSave() {
         if (this.formData) {
-            this.taskService.editTask(this.taskId, this.transformTaskToApiObject(this.formData)).subscribe(
-                (response) => {
+            this.taskService.editTask(this.taskId, this.transformTaskToApiObject(this.formData))
+                .subscribe((response) => {
                     this.notificationService.openSuccessNotification('success.edit');
                     this.router.navigate(['tasks/list']);
                 }, (error) => {
@@ -67,8 +77,30 @@ export class TaskEditComponent implements OnInit {
         }
         const taskComment: TaskComment = {
             description: this.addCommentForm.value.newComment.toString(),
-            taskId: this.taskId
+            taskId: this.taskId,
+            type: 'TEXT'
         };
+
+        this.onSendCommentService(taskComment);
+    }
+
+    public onAttachmentAdded() {
+        const taskComment: TaskComment = {
+            description: '',
+            taskId: this.taskId,
+            type: 'ATTACHMENT',
+            attachment: {
+                type: 'COMMENT',
+                format: this.attachmentFormat,
+                fileName: this.attachmentFileName,
+                filePath: this.attachmentPathName
+            }
+        };
+
+        this.onSendCommentService(taskComment);
+    }
+
+    public onSendCommentService(taskComment) {
         this.loading = true;
         this.taskCommentService.addComment(taskComment)
             .subscribe((commentResponse: TaskCommentResponse) => {
@@ -84,15 +116,84 @@ export class TaskEditComponent implements OnInit {
     public getAllComments() {
         this.loading = true;
         this.taskCommentService.getAllComment(this.taskId)
-            .pipe(
-                tap(() => this.loading = false)
-            )
+            .pipe(tap(() => this.loading = false))
             .subscribe((comments: TaskCommentResponse[]) => {
                 this.comments = [...comments].reverse();
             }, (error) => {
                 this.notificationService.openErrorNotification('error.loadComments');
             });
 
+    }
+
+    public hasRoleUploadImage() {
+        return this.authService.hasRole(Role.RoleUploadImage);
+    }
+
+    public formDataChange($event) {
+        setTimeout(() => {
+            this.formData = $event;
+        }, 200);
+    }
+
+    public onFileChange(event) {
+        const file = event.target.files[0];
+        if (!file) {
+            return;
+        }
+        const reader = new FileReader();
+        reader.onload = () => {
+            this.imagesService.uploadImages([file])
+                .subscribe((data) => {
+                    this.attachmentFormat = file.name.split('.')
+                                                .pop()
+                                                .toUpperCase();
+                    this.attachmentFileName = file.name;
+                    this.attachmentPathName = data.fileNames[file.name].replace(/^.*[\\\/]/, '');
+                    this.onAttachmentAdded();
+                }, () => {
+                    this.attachmentFormat = '';
+                    this.attachmentFileName = '';
+                    this.attachmentPathName = '';
+                    this.notificationService.openErrorNotification('error.imageUpload');
+                });
+        };
+        reader.readAsDataURL(file);
+    }
+
+    public allowReadComment(): boolean {
+        return this.hasRoleReadComment() || this.hasRoleReadCommentInAssignProject();
+    }
+
+    public allowUpdateTask(): boolean {
+        return this.hasRoleUpdateTask() || this.hasRoleUpdateTaskInAssignProject();
+    }
+
+    public allowCreateComment(): boolean {
+        return this.hasRoleCreateComment() || this.hasRoleCreateCommentInAssignProject();
+    }
+
+    private hasRoleUpdateTask(): boolean {
+        return this.authService.hasRole(Role.RoleUpdateTask);
+    }
+
+    private hasRoleUpdateTaskInAssignProject(): boolean {
+        return this.authService.hasRole(Role.RoleUpdateTaskInAssignProject);
+    }
+
+    private hasRoleReadComment(): boolean {
+        return this.authService.hasRole(Role.RoleReadComment);
+    }
+
+    private hasRoleReadCommentInAssignProject(): boolean {
+        return this.authService.hasRole(Role.RoleReadCommentInAssignProject);
+    }
+
+    private hasRoleCreateComment(): boolean {
+        return this.authService.hasRole(Role.RoleCreateComment);
+    }
+
+    private hasRoleCreateCommentInAssignProject(): boolean {
+        return this.authService.hasRole(Role.RoleCreateCommentInAssignProject);
     }
 
     private transformTaskToApiObject(formObject: any): any {
@@ -111,12 +212,6 @@ export class TaskEditComponent implements OnInit {
             apiObject.trafficLight = formObject.trafficLight;
         }
         return apiObject;
-    }
-
-    public formDataChange($event) {
-        setTimeout(() => {
-            this.formData = $event;
-        }, 200);
     }
 
 }
