@@ -5,9 +5,10 @@ import { MaterialModule } from '../hazlenut/hazelnut-common';
 import { SmallComponentsModule } from '../hazlenut/small-components';
 import { NotificationSnackBarComponent } from '../hazlenut/small-components/notifications';
 import { PdfDialogComponent } from './dialog/pdf-dialog/pdf-dialog.component';
+import { ImageDialogComponent } from './dialog/image-dialog/image-dialog.component';
 
 @NgModule({
-    declarations: [PdfDialogComponent],
+    declarations: [PdfDialogComponent, ImageDialogComponent],
     imports: [
         CommonModule,
         SmallComponentsModule,
@@ -16,7 +17,8 @@ import { PdfDialogComponent } from './dialog/pdf-dialog/pdf-dialog.component';
     ],
     exports: [
         NotificationSnackBarComponent,
-        PdfDialogComponent
+        PdfDialogComponent,
+        ImageDialogComponent
     ]
 })
 export class ComponentsModule {
