@@ -1,16 +1,14 @@
-import { HttpClient, HttpParams } from '@angular/common/http';
+import { HttpClient, HttpParams, HttpResponse } from '@angular/common/http';
 import { Inject } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
+import { ReportType } from '../../../enums/report-type.enum';
 // TODO: abstract service shouldn't be dependent od core-table
 import { TableChangeEvent } from '../../core-table/table-change-event';
-import {
-    NOTIFICATION_WRAPPER_TOKEN,
-    NotificationWrapper
-} from '../../small-components/notifications/notification.wrapper';
-import { HazelnutConfig } from '../config/hazelnut-config';
+import { NOTIFICATION_WRAPPER_TOKEN, NotificationWrapper } from '../../small-components/notifications/notification.wrapper';
+import { hazelnutConfig } from '../config/hazelnut-config';
 import { StringMap } from '../hazelnut';
-import { BrowseResponse, Direction, Filter, PostContent, Property, Sort } from '../models';
+import { BrowseResponse, Direction, Filter, PostContent, Sort } from '../models';
 import { CountModel } from '../models/count.model';
 import { AbstractServiceParams, CoreService } from './core-service.service';
 
@@ -18,9 +16,7 @@ import { AbstractServiceParams, CoreService } from './core-service.service';
  * HTTP methods which find usage in multiple projects are here
  */
 export abstract class AbstractService<T = any> extends CoreService<T> {
-    protected constructor(http: HttpClient,
-                          @Inject(NOTIFICATION_WRAPPER_TOKEN) notificationService: NotificationWrapper,
-                          protected readonly urlKey: string) {
+    protected constructor(http: HttpClient, @Inject(NOTIFICATION_WRAPPER_TOKEN) notificationService: NotificationWrapper, protected readonly urlKey: string) {
         super(http, notificationService);
     }
 
@@ -28,15 +24,18 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
      * Function return results from browse api for {@link CoreTableComponent}
      */
     public browseTable(params: TableChangeEvent, url = this.urlKey + '/browse'): Observable<BrowseResponse<T>> {
-        const filters: Filter[] = Object.keys(params.filters).map((key) => params.filters[key]);
+        const filters: Filter[] = Object.keys(params.filters)
+                                        .map((key: string) => params.filters[key]);
         const sorts: Sort[] = [];
 
         if (params.sortActive) {
             sorts.push(new Sort(params.sortActive, params.sortDirection as Direction));
         }
-        return this.browseInner(`${HazelnutConfig.URL_API}/${url}`,
+
+        return this.browseInner(`${hazelnutConfig.URL_API}/${url}`,
             PostContent.create(params.pageSize, params.pageSize * params.pageIndex, filters, sorts),
-            (response) => new BrowseResponse(this.extractListData(response), response.totalElements));
+            (response: any) => new BrowseResponse(this.extractListData(response), response.totalElements)
+        );
     }
 
     /**
@@ -64,9 +63,10 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
      */
     public getDetail(id: number | string, params?: StringMap): Observable<T> {
         const realId = id ? `/${id}` : '';
+
         return this.get({
             params,
-            url: `${HazelnutConfig.URL_API}/${this.urlKey}${realId}`,
+            url: `${hazelnutConfig.URL_API}/${this.urlKey}${realId}`,
             mapFunction: this.extractDetail
         });
     }
@@ -79,7 +79,7 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
     public add(body: T, additionalUrl = ''): Observable<T> {
         return this.post({
             body,
-            url: `${HazelnutConfig.URL_API}/${this.urlKey}${additionalUrl}`,
+            url: `${hazelnutConfig.URL_API}/${this.urlKey}${additionalUrl}`,
             mapFunction: this.extractDetail,
         });
     }
@@ -93,16 +93,17 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
     public update(id: number, body: T): Observable<T> {
         return this.put({
             body,
-            url: `${HazelnutConfig.URL_API}/${this.urlKey}/${id}`,
+            url: `${hazelnutConfig.URL_API}/${this.urlKey}/${id}`,
             mapFunction: this.extractDetail
         });
     }
 
     public deleteById(id: number, params?: StringMap): Observable<any> {
         const realId = id ? `/${id}` : '';
+
         return this.delete({
             params,
-            url: `${HazelnutConfig.URL_API}/${this.urlKey}${realId}`,
+            url: `${hazelnutConfig.URL_API}/${this.urlKey}${realId}`,
             mapFunction: this.extractDetail
         });
     }
@@ -110,6 +111,7 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
     protected uploadFile<S = T>(url: string, file: File, fileName: string, mapFunction: AbstractServiceParams<S>['mapFunction']): Observable<S> {
         const fd = new FormData();
         fd.append(fileName, file, file ? file.name : undefined);
+
         return this.post({
             url,
             mapFunction,
@@ -120,17 +122,12 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
     /**
      *
      */
-    protected browse<S = T>(filter: Filter[] = [],
-                            sort: Sort[] = [new Sort()],
-                            limit: number = HazelnutConfig.BROWSE_LIMIT,
-                            offset = 0): Observable<T[]> {
+    protected browse<S = T>(filter: Filter[] = [], sort: Sort[] = [new Sort()], limit: number = hazelnutConfig.BROWSE_LIMIT, offset = 0): Observable<T[]> {
         if (sort.length === 0) {
             sort.push(new Sort());
         }
 
-        return this.browseInner(`${HazelnutConfig.URL_API}/${this.urlKey}/browse`,
-            PostContent.create(limit, offset, filter, sort),
-            this.extractListData);
+        return this.browseInner(`${hazelnutConfig.URL_API}/${this.urlKey}/browse`, PostContent.create(limit, offset, filter, sort), this.extractListData);
     }
 
     protected browseWithSummary<S = T>(postContent: PostContent, additionalUrl = ''): Observable<BrowseResponse<S>> {
@@ -149,9 +146,8 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
         } else if (!this.urlKey.endsWith('/')) {
             additionalUrl = '/' + additionalUrl;
         }
-        return this.browseInner<BrowseResponse<S>>(`${HazelnutConfig.URL_API}/${this.urlKey}${additionalUrl}browse`,
-            postContent,
-            this.extractDetail);
+
+        return this.browseInner<BrowseResponse<S>>(`${hazelnutConfig.URL_API}/${this.urlKey}${additionalUrl}browse`, postContent, this.extractDetail);
     }
 
     /**
@@ -163,8 +159,9 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
                 criteria: filter,
             },
         };
+
         return this.post({
-            url: `${HazelnutConfig.URL_API}/${this.urlKey}${infix}/count`,
+            url: `${hazelnutConfig.URL_API}/${this.urlKey}${infix}/count`,
             body: content,
             mapFunction: this.extractDetail
         });
@@ -189,7 +186,7 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
 
         return this.post<S[]>({
             body,
-            url: `${HazelnutConfig.URL_API}/${this.urlKey}/filter`,
+            url: `${hazelnutConfig.URL_API}/${this.urlKey}/filter`,
             mapFunction: this.extractListData,
         });
     }
@@ -198,7 +195,7 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
      * Function extract all data from response
      *
      */
-    protected extractListData<S = T>(res: { content: S[] }): S[] {
+    protected extractListData<S = T>(res: {content: S[]}): S[] {
         return (res && res.content) || [];
     }
 
@@ -217,7 +214,7 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
         return this.post({
             url,
             mapFunction,
-            body: content.prepareAndGet(HazelnutConfig.BROWSE_LIMIT)
+            body: content.prepareAndGet(hazelnutConfig.BROWSE_LIMIT)
         });
     }
 
@@ -227,12 +224,16 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
      * @param {Sort[]} sort
      * @returns {Observable<T>}
      */
-    protected report<S = T>(filter: Filter[] = [], sort: Sort[] = [new Sort()]): Observable<any> {
+    protected report<S = T>(filter: Filter[] = [], sort: Sort[] = [new Sort()], projectId: number): Observable<any> {
         if (sort && sort.length === 0) {
             sort.push(new Sort());
         }
-        const content = {filterCriteria: {criteria: filter}, sortingCriteria: {criteria: sort}};
-        return this.postBlob(`${HazelnutConfig.URL_API}/${this.urlKey}/report`, content, this.extractDetail);
+        const content = {
+            filterCriteria: {criteria: filter},
+            sortingCriteria: {criteria: sort}
+        };
+
+        return this.postBlob(`${hazelnutConfig.URL_API}/${this.urlKey}/project/${projectId}/report`, content, this.extractDetail);
     }
 
     /**
@@ -242,19 +243,14 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
      * @param params
      * @returns {Observable<S>}
      */
-    protected postBlob<S = T[]>(url: string,
-                                body: any,
-                                mapFunction: (response: any) => S,
-                                params: HttpParams | { [param: string]: string | string[]; } = {}): Observable<S> {
+    protected postBlob<S = T[]>(url: string, body: any, mapFunction: (response: any) => S, params: HttpParams | {[param: string]: string | string[]; } = {}): Observable<S> {
         return this.http.post(url, body, {
-            params,
-            headers: this.getHeader(),
-            responseType: 'blob',
-            observe: 'response'
-        }).pipe(
-            map(mapFunction),
-            catchError(this.handleError),
-        );
+                       params,
+                       headers: this.getHeader(),
+                       responseType: 'blob',
+                       observe: 'response'
+                   })
+                   .pipe(map(mapFunction), catchError(this.handleError),);
     }
 
     /**
@@ -264,7 +260,7 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
      * @returns {Observable<T>}
      */
     protected reportGet<S = T>(projectId: number, reportId: number): Observable<any> {
-        return this.getBlob(`${HazelnutConfig.URL_API}/report/${projectId}/${reportId}`, this.extractDetail);
+        return this.getBlob(`${hazelnutConfig.URL_API}/report/${reportId === 1 ? ReportType.ToDoList : ReportType.RedFlagList}/project/${projectId}`, this.extractDetail);
     }
 
     /**
@@ -274,18 +270,14 @@ export abstract class AbstractService<T = any> extends CoreService<T> {
      * @param params
      * @returns {Observable<S>}
      */
-    protected getBlob<S = T[]>(url: string,
-                               mapFunction: (response: any) => S,
-                               params: HttpParams | { [param: string]: string | string[]; } = {}): Observable<S> {
+    protected getBlob<S = T[]>(url: string, mapFunction: (response: any) => S, params: HttpParams | {[param: string]: string | string[]; } = {}): Observable<S> {
         return this.http.get(url, {
-            params,
-            headers: this.getHeader(),
-            responseType: 'blob',
-            observe: 'response'
-        }).pipe(
-            map(mapFunction),
-            catchError(this.handleError),
-        );
+                       params,
+                       headers: this.getHeader(),
+                       responseType: 'blob',
+                       observe: 'response'
+                   })
+                   .pipe(map(mapFunction), catchError(this.handleError),);
     }
 
 }
